tests/requirements-test.txt
# tests/requirements-test.txt
# Dépendances pour les tests unitaires

pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0